+++ docs/pages/_en/beta/reference/release-notes.md
@@ -10,13 +10,16 @@ Some of the changes mentioned in sections below, might break your existing Grass
 {% include ltr/warning_note.html note=breaking_changes_notes %}
 
 <!-- most recent release should be on top -->
-{% include ltr/release-header.html version="0.5" time="04/21/2021" %}
+{% include ltr/release-header.html version="0.5" time="04/22/2021" %}
 
 ### New Features
 * Now 'Add Component (Location)' tries to reuse previous iteration element.
 
 ### Fixes
 * Now Transaction warnings and errors are displayed in the component balloon only.
+
+### API
+* Removed Obsolete `Revit.ApplicationUI` and `Revit.CurrentUsersDataFolderPath` properties.
 
 {% include ltr/release-header.html version="0.4" time="04/20/2021" %}